requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0